Refactor the TestNG Transformer directory structure
