+++ autotest/OldTests/src/test/java/testng/FailureRetryAnalyzer.java
@@ -1,9 +1,10 @@
-package retry;
+package testng;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.testng.IRetryAnalyzer;
 import org.testng.ITestResult;
+import testng.annotation.RetryTest;
 
 public class FailureRetryAnalyzer implements IRetryAnalyzer {
   int currentRetry = 0;
